# swagger-spec.json
{"openapi":"3.0.0","paths":{"/api":{"get":{"operationId":"getHello","parameters":[],"responses":{"200":{"description":""}}}},"/api/role":{"post":{"operationId":"create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRoleDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoleEntity"}}}}},"tags":["roles"]},"get":{"operationId":"findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RoleEntity"}}}}}},"tags":["roles"]}},"/api/role/page":{"get":{"operationId":"pagination","parameters":[{"name":"page","required":false,"in":"query","example":1,"schema":{"type":"number"}},{"name":"perPage","required":false,"in":"query","example":10,"schema":{"type":"number"}},{"name":"orderBy","required":false,"in":"query","example":{"id":"desc"},"schema":{"type":"object"}},{"name":"where","required":false,"in":"query","example":{"id":"4"},"schema":{"type":"object"}}],"responses":{"200":{"description":""}},"tags":["roles"]}},"/api/role/{id}":{"get":{"operationId":"findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoleEntity"}}}}},"tags":["roles"]},"patch":{"operationId":"update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRoleDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoleEntity"}}}}},"tags":["roles"]},"delete":{"operationId":"remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoleEntity"}}}}},"tags":["roles"]}},"/api/user":{"post":{"operationId":"create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}}},"tags":["users"]},"get":{"operationId":"findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserEntity"}}}}}},"tags":["users"]}},"/api/user/page":{"get":{"operationId":"pagination","parameters":[{"name":"page","required":false,"in":"query","example":1,"schema":{"type":"number"}},{"name":"perPage","required":false,"in":"query","example":10,"schema":{"type":"number"}},{"name":"orderBy","required":false,"in":"query","example":{"id":"desc"},"schema":{"type":"object"}},{"name":"where","required":false,"in":"query","example":{"id":"4"},"schema":{"type":"object"}}],"responses":{"200":{"description":""}},"tags":["users"]}},"/api/user/login":{"post":{"operationId":"login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":""}},"tags":["users"]}},"/api/user/{id}":{"get":{"operationId":"findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}}},"tags":["users"]},"patch":{"operationId":"update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}}},"tags":["users"]},"delete":{"operationId":"remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserEntity"}}}}},"tags":["users"]}},"/api/product":{"post":{"operationId":"create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProductDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductEntity"}}}}},"tags":["products"]},"get":{"operationId":"findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductEntity"}}}}}},"tags":["products"]}},"/api/product/page":{"get":{"operationId":"pagination","parameters":[{"name":"page","required":false,"in":"query","example":1,"schema":{"type":"number"}},{"name":"perPage","required":false,"in":"query","example":10,"schema":{"type":"number"}},{"name":"orderBy","required":false,"in":"query","example":{"id":"desc"},"schema":{"type":"object"}},{"name":"where","required":false,"in":"query","example":{"id":"4"},"schema":{"type":"object"}}],"responses":{"200":{"description":""}},"tags":["products"]}},"/api/product/{id}":{"get":{"operationId":"findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductEntity"}}}}},"tags":["products"]},"patch":{"operationId":"update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProductDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductEntity"}}}}},"tags":["products"]},"delete":{"operationId":"remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductEntity"}}}}},"tags":["products"]}},"/api/order":{"post":{"operationId":"create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateOrderDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderEntity"}}}}},"tags":["orders"]},"get":{"operationId":"findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OrderEntity"}}}}}},"tags":["orders"]}},"/api/order/page":{"get":{"operationId":"pagination","parameters":[{"name":"page","required":false,"in":"query","example":1,"schema":{"type":"number"}},{"name":"perPage","required":false,"in":"query","example":10,"schema":{"type":"number"}},{"name":"orderBy","required":false,"in":"query","example":{"id":"desc"},"schema":{"type":"object"}},{"name":"where","required":false,"in":"query","example":{"id":"4"},"schema":{"type":"object"}}],"responses":{"200":{"description":""}},"tags":["orders"]}},"/api/order/{id}":{"get":{"operationId":"findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderEntity"}}}}},"tags":["orders"]},"patch":{"operationId":"update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateOrderDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderEntity"}}}}},"tags":["orders"]},"delete":{"operationId":"remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderEntity"}}}}},"tags":["orders"]}}},"info":{"title":"Operator Hub","description":"App","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"access-token":{"scheme":"Bearer","bearerFormat":"Token","type":"http","in":"header"}},"schemas":{"CreateRoleDto":{"type":"object","properties":{"name":{"type":"string"},"default":{"type":"boolean","default":false}},"required":["name"]},"RoleEntity":{"type":"object","properties":{"id":{"type":"number"},"deleted":{"format":"date-time","type":"string","default":false},"name":{"type":"string"},"default":{"type":"boolean","default":false}},"required":["id","name"]},"UpdateRoleDto":{"type":"object","properties":{"name":{"type":"string"},"default":{"type":"boolean","default":false}}},"CreateUserDto":{"type":"object","properties":{"email":{"type":"string"},"username":{"type":"string"},"password":{"type":"string"}},"required":["email","username","password"]},"UserEntity":{"type":"object","properties":{"id":{"type":"number"},"deleted":{"format":"date-time","type":"string","default":false},"email":{"type":"string"},"username":{"type":"string"},"password":{"type":"string"}},"required":["id","email","username","password"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"UpdateUserDto":{"type":"object","properties":{"email":{"type":"string"},"username":{"type":"string"},"password":{"type":"string"}}},"CreateProductDto":{"type":"object","properties":{"durationTime":{"type":"number"},"name":{"type":"string"},"userId":{"type":"number"}},"required":["durationTime","name","userId"]},"ProductEntity":{"type":"object","properties":{"id":{"type":"number"},"deleted":{"format":"date-time","type":"string"},"durationTime":{"type":"number"},"name":{"type":"string"},"userId":{"type":"number"},"assignee":{"$ref":"#/components/schemas/UserEntity"}},"required":["id","deleted","durationTime","name","userId","assignee"]},"UpdateProductDto":{"type":"object","properties":{"durationTime":{"type":"number"},"name":{"type":"string"},"userId":{"type":"number"}}},"CreateOrderDto":{"type":"object","properties":{"productId":{"type":"number"},"userId":{"type":"number"}},"required":["productId","userId"]},"OrderEntity":{"type":"object","properties":{"id":{"type":"number"},"orderNumber":{"type":"string"},"orderTime":{"format":"date-time","type":"string","default":false},"dueTime":{"format":"date-time","type":"string","default":false},"deleted":{"format":"date-time","type":"string","default":false},"productId":{"type":"number"},"userId":{"type":"number"}},"required":["id","orderNumber","productId","userId"]},"UpdateOrderDto":{"type":"object","properties":{"productId":{"type":"number"},"userId":{"type":"number"}}}}}}